+++ src/Model/Response/Component/Element/Element.php
@@ -6,4 +6,11 @@ namespace Wingu\Engine\SDK\Model\Response\Component\Element;
 
 interface Element
 {
+    public function name() : string;
+
+    public function label() : string;
+
+    public function required() : bool;
+
+    public function persistent() : bool;
 }