+++ tests/Api/Channel/Geofence/GeofenceApiTest.php
@@ -29,9 +29,6 @@ use Wingu\Engine\SDK\Model\Response\Component\ContactAddress;
 use Wingu\Engine\SDK\Model\Response\Component\ContactExternalLinks;
 use Wingu\Engine\SDK\Model\Response\Component\Coupon;
 use Wingu\Engine\SDK\Model\Response\Component\CouponBarcode;
-use Wingu\Engine\SDK\Model\Response\Component\Element\Input;
-use Wingu\Engine\SDK\Model\Response\Component\Element\Select;
-use Wingu\Engine\SDK\Model\Response\Component\Element\SelectOption;
 use Wingu\Engine\SDK\Model\Response\Component\Files;
 use Wingu\Engine\SDK\Model\Response\Component\FilesFile as File;
 use Wingu\Engine\SDK\Model\Response\Component\Image;
@@ -41,12 +38,9 @@ use Wingu\Engine\SDK\Model\Response\Component\ImageGalleryImage as OuterImage;
 use Wingu\Engine\SDK\Model\Response\Component\ImageMetadata;
 use Wingu\Engine\SDK\Model\Response\Component\ImageMetadata as Metadata;
 use Wingu\Engine\SDK\Model\Response\Component\Location;
-use Wingu\Engine\SDK\Model\Response\Component\PrivateForm;
 use Wingu\Engine\SDK\Model\Response\Component\PrivateWebhook;
 use Wingu\Engine\SDK\Model\Response\Component\Rating;
 use Wingu\Engine\SDK\Model\Response\Component\Separator;
-use Wingu\Engine\SDK\Model\Response\Component\SubmitDestination\Email;
-use Wingu\Engine\SDK\Model\Response\Component\SubmitDestination\Endpoint;
 use Wingu\Engine\SDK\Model\Response\Component\SurveyMonkey;
 use Wingu\Engine\SDK\Model\Response\Component\Video;
 use Wingu\Engine\SDK\Model\Response\Content\Deck;
@@ -56,9 +50,12 @@ use Wingu\Engine\SDK\Model\Response\Content\PrivateContent;
 use Wingu\Engine\SDK\Model\Response\Content\PrivateListContent;
 use Wingu\Engine\SDK\Model\Response\Coordinates;
 use Wingu\Engine\SDK\Tests\Api\ChannelApiTestCase;
+use Wingu\Engine\SDK\Tests\Api\Expected\Loader\PrivateComponent;
 
 final class GeofenceApiTest extends ChannelApiTestCase
 {
+    use PrivateComponent;
+
     public function testGeofenceReturnsPublicGeofence() : void
     {
         $configurationMock = new Configuration();
@@ -252,48 +249,7 @@ final class GeofenceApiTest extends ChannelApiTestCase
                                 new Card(
                                     'e68e8946-6dcc-48bd-bc8f-f252f9dd1c63',
                                     new Position(8),
-                                    new PrivateForm(
-                                        'a787b433-7dfb-4e59-b26a-6786e6fd44ac',
-                                        new \DateTime('2018-05-18T08:22:41+0000'),
-                                        'Form component survey',
-                                        [
-                                            new Input('full_name', 'Your name', true, 'text'),
-                                            new Input('birthday', 'Birthday', false, 'date'),
-                                            new Select(
-                                                'gender',
-                                                'Gender',
-                                                false,
-                                                false,
-                                                [
-                                                    new SelectOption('Male', 'm'),
-                                                    new SelectOption('Female', 'f'),
-                                                ]
-                                            ),
-                                            new Select(
-                                                'dessert',
-                                                'Dessert',
-                                                true,
-                                                true,
-                                                [
-                                                    new SelectOption('Jello', 'jello'),
-                                                    new SelectOption('Apple pie', 'apple_pie'),
-                                                    new SelectOption('Schnitzel', 'schnitzel'),
-                                                ]
-                                            ),
-                                            new Input('text', 'Element text', false, 'text'),
-                                            new Input('textarea', 'Element textarea', false, 'textarea'),
-                                            new Input('email', 'Element email', false, 'email'),
-                                            new Input('url', 'Element url', false, 'url'),
-                                            new Input('date', 'Element date', false, 'date'),
-                                            new Input('datetime', 'Element datetime', false, 'datetime'),
-                                            new Input('time', 'Element time', false, 'time'),
-                                        ],
-                                        [
-                                            new Email('test+form-component@wingu.de'),
-                                            new Endpoint('https://httpbin.org/status/200', []),
-                                        ],
-                                        'Thank you for your feedback!'
-                                    )
+                                    $this->getExpectedFormComponent()
                                 ),
                                 new Card(
                                     '4c680d50-c9c8-4794-be8d-70b45021eb86',